refactor: product 패키지 정리
